gitreaper: Add provision to change sync direction
We generally sync from upstream to downstream. But when you manage
both upstream and downstream repos, you might want to sync from downstream
to upstream. Add provision for that. Since it is an unusual request
and the effects are far more serious, confirm with the user once before
syncing. The user has to type YES(all caps) when prompted.

Signed-off-by: Vijai Kumar K <vijaikumar.kanagarajan@gmail.com>